Move prompt serializer to prompts/api
